Implementation of `limited` tracer state in which most spans should not be created

+++ src/DDTrace/Integrations/Memcached/MemcachedIntegration.php
@@ -60,70 +60,123 @@ class MemcachedIntegration extends Integration
 
         // bool Memcached::add ( string $key , mixed $value [, int $expiration ] )
         dd_trace('Memcached', 'add', function () {
+            if (GlobalTracer::get()->limited()) {
+                return dd_trace_forward_call();
+            }
+
             return MemcachedIntegration::traceCommand($this, 'add', func_get_args());
         });
 
         // bool Memcached::addByKey ( string $server_key , string $key , mixed $value [, int $expiration ] )
         dd_trace('Memcached', 'addByKey', function () {
+            if (GlobalTracer::get()->limited()) {
+                return dd_trace_forward_call();
+            }
+
             return MemcachedIntegration::traceCommandByKey($this, 'addByKey', func_get_args());
         });
 
         // bool Memcached::append ( string $key , string $value )
         dd_trace('Memcached', 'append', function () {
+            if (GlobalTracer::get()->limited()) {
+                return dd_trace_forward_call();
+            }
+
             return MemcachedIntegration::traceCommand($this, 'append', func_get_args());
         });
 
         // bool Memcached::appendByKey ( string $server_key , string $key , string $value )
         dd_trace('Memcached', 'appendByKey', function () {
+            if (GlobalTracer::get()->limited()) {
+                return dd_trace_forward_call();
+            }
+
             return MemcachedIntegration::traceCommandByKey($this, 'appendByKey', func_get_args());
         });
 
         // bool Memcached::cas ( float $cas_token , string $key , mixed $value [, int $expiration ] )
         dd_trace('Memcached', 'cas', function () {
+            if (GlobalTracer::get()->limited()) {
+                return dd_trace_forward_call();
+            }
+
             return MemcachedIntegration::traceCas($this, func_get_args());
         });
 
         // bool Memcached::casByKey ( float $cas_token , string $server_key , string $key , mixed $value
         //     [, int $expiration ] )
         dd_trace('Memcached', 'casByKey', function () {
+            if (GlobalTracer::get()->limited()) {
+                return dd_trace_forward_call();
+            }
+
             return MemcachedIntegration::traceCasByKey($this, func_get_args());
         });
 
         // int Memcached::decrement ( string $key [, int $offset = 1 [, int $initial_value = 0
         //     [, int $expiry = 0 ]]] )
         dd_trace('Memcached', 'decrement', function () {
+            if (GlobalTracer::get()->limited()) {
+                return dd_trace_forward_call();
+            }
+
             return MemcachedIntegration::traceCommand($this, 'decrement', func_get_args());
         });
 
         // int Memcached::decrementByKey ( string $server_key , string $key
         //      [, int $offset = 1 [, int $initial_value = 0 [, int $expiry = 0 ]]] )
         dd_trace('Memcached', 'decrementByKey', function () {
+            if (GlobalTracer::get()->limited()) {
+                return dd_trace_forward_call();
+            }
+
             return MemcachedIntegration::traceCommandByKey($this, 'decrementByKey', func_get_args());
         });
 
         // bool Memcached::delete ( string $key [, int $time = 0 ] )
         dd_trace('Memcached', 'delete', function () {
+            if (GlobalTracer::get()->limited()) {
+                return dd_trace_forward_call();
+            }
+
             return MemcachedIntegration::traceCommand($this, 'delete', func_get_args());
         });
 
         // bool Memcached::deleteByKey ( string $server_key , string $key [, int $time = 0 ] )
         dd_trace('Memcached', 'deleteByKey', function () {
+            if (GlobalTracer::get()->limited()) {
+                return dd_trace_forward_call();
+            }
+
             return MemcachedIntegration::traceCommandByKey($this, 'deleteByKey', func_get_args());
         });
 
         // array Memcached::deleteMulti ( array $keys [, int $time = 0 ] )
         dd_trace('Memcached', 'deleteMulti', function () {
+            if (GlobalTracer::get()->limited()) {
+                return dd_trace_forward_call();
+            }
+
             return MemcachedIntegration::traceMulti($this, 'deleteMulti', func_get_args());
         });
 
         // bool Memcached::deleteMultiByKey ( string $server_key , array $keys [, int $time = 0 ] )
         dd_trace('Memcached', 'deleteMultiByKey', function () {
+            if (GlobalTracer::get()->limited()) {
+                return dd_trace_forward_call();
+            }
+
             return MemcachedIntegration::traceMultiByKey($this, 'deleteMultiByKey', func_get_args());
         });
 
         // bool Memcached::flush ([ int $delay = 0 ] )
         dd_trace('Memcached', 'flush', function () {
-            $scope = GlobalTracer::get()->startIntegrationScopeAndSpan(
+            $tracer = GlobalTracer::get();
+            if ($tracer->limited()) {
+                return dd_trace_forward_call();
+            }
+
+            $scope = $tracer->startIntegrationScopeAndSpan(
                 MemcachedIntegration::getInstance(),
                 'Memcached.flush'
             );
@@ -138,83 +191,147 @@ class MemcachedIntegration extends Integration
 
         // mixed Memcached::get ( string $key [, callable $cache_cb [, int &$flags ]] )
         dd_trace('Memcached', 'get', function () {
+            if (GlobalTracer::get()->limited()) {
+                return dd_trace_forward_call();
+            }
+
             return MemcachedIntegration::traceCommand($this, 'get', func_get_args());
         });
 
         // mixed Memcached::getByKey ( string $server_key , string $key [, callable $cache_cb [, int $flags ]] )
         dd_trace('Memcached', 'getByKey', function () {
+            if (GlobalTracer::get()->limited()) {
+                return dd_trace_forward_call();
+            }
+
             return MemcachedIntegration::traceCommandByKey($this, 'getByKey', func_get_args());
         });
 
         // mixed Memcached::getMulti ( array $keys [, int $flags ] )
         dd_trace('Memcached', 'getMulti', function () {
+            if (GlobalTracer::get()->limited()) {
+                return dd_trace_forward_call();
+            }
+
             return MemcachedIntegration::traceMulti($this, 'getMulti', func_get_args());
         });
 
         // array Memcached::getMultiByKey ( string $server_key , array $keys [, int $flags ] )
         dd_trace('Memcached', 'getMultiByKey', function () {
+            if (GlobalTracer::get()->limited()) {
+                return dd_trace_forward_call();
+            }
+
             return MemcachedIntegration::traceMultiByKey($this, 'getMultiByKey', func_get_args());
         });
 
         // int Memcached::increment ( string $key [, int $offset = 1 [, int $initial_value = 0
         //     [, int $expiry = 0 ]]] )
         dd_trace('Memcached', 'increment', function () {
+            if (GlobalTracer::get()->limited()) {
+                return dd_trace_forward_call();
+            }
+
             return MemcachedIntegration::traceCommand($this, 'increment', func_get_args());
         });
 
         // int Memcached::incrementByKey ( string $server_key , string $key [, int $offset = 1
         //     [, int $initial_value = 0 [, int $expiry = 0 ]]] )
         dd_trace('Memcached', 'incrementByKey', function () {
+            if (GlobalTracer::get()->limited()) {
+                return dd_trace_forward_call();
+            }
+
             return MemcachedIntegration::traceCommandByKey($this, 'incrementByKey', func_get_args());
         });
 
         // bool Memcached::prepend ( string $key , string $value )
         dd_trace('Memcached', 'prepend', function () {
+            if (GlobalTracer::get()->limited()) {
+                return dd_trace_forward_call();
+            }
+
             return MemcachedIntegration::traceCommand($this, 'prepend', func_get_args());
         });
 
         // bool Memcached::prependByKey ( string $server_key , string $key , string $value )
         dd_trace('Memcached', 'prependByKey', function () {
+            if (GlobalTracer::get()->limited()) {
+                return dd_trace_forward_call();
+            }
+
             return MemcachedIntegration::traceCommandByKey($this, 'prependByKey', func_get_args());
         });
 
         // bool Memcached::replace ( string $key , mixed $value [, int $expiration ] )
         dd_trace('Memcached', 'replace', function () {
+            if (GlobalTracer::get()->limited()) {
+                return dd_trace_forward_call();
+            }
+
             return MemcachedIntegration::traceCommand($this, 'replace', func_get_args());
         });
 
         // bool Memcached::replaceByKey ( string $server_key , string $key , mixed $value [, int $expiration  ] )
         dd_trace('Memcached', 'replaceByKey', function () {
+            if (GlobalTracer::get()->limited()) {
+                return dd_trace_forward_call();
+            }
+
             return MemcachedIntegration::traceCommandByKey($this, 'replaceByKey', func_get_args());
         });
 
         // bool Memcached::set ( string $key , mixed $value [, int $expiration ] )
         dd_trace('Memcached', 'set', function () {
+            if (GlobalTracer::get()->limited()) {
+                return dd_trace_forward_call();
+            }
+
             return MemcachedIntegration::traceCommand($this, 'set', func_get_args());
         });
 
         // bool Memcached::setByKey ( string $server_key , string $key , mixed $value [, int $expiration ] )
         dd_trace('Memcached', 'setByKey', function () {
+            if (GlobalTracer::get()->limited()) {
+                return dd_trace_forward_call();
+            }
+
             return MemcachedIntegration::traceCommandByKey($this, 'setByKey', func_get_args());
         });
 
         // bool Memcached::setMulti ( array $items [, int $expiration ] )
         dd_trace('Memcached', 'setMulti', function () {
+            if (GlobalTracer::get()->limited()) {
+                return dd_trace_forward_call();
+            }
+
             return MemcachedIntegration::traceMulti($this, 'setMulti', func_get_args());
         });
 
         // bool Memcached::setMultiByKey ( string $server_key , array $items [, int $expiration ] )
         dd_trace('Memcached', 'setMultiByKey', function () {
+            if (GlobalTracer::get()->limited()) {
+                return dd_trace_forward_call();
+            }
+
             return MemcachedIntegration::traceMultiByKey($this, 'setMultiByKey', func_get_args());
         });
 
         // bool Memcached::touch ( string $key , int $expiration )
         dd_trace('Memcached', 'touch', function () {
+            if (GlobalTracer::get()->limited()) {
+                return dd_trace_forward_call();
+            }
+
             return MemcachedIntegration::traceCommand($this, 'touch', func_get_args());
         });
 
         // bool Memcached::touchByKey ( string $server_key , string $key , int $expiration )
         dd_trace('Memcached', 'touchByKey', function () {
+            if (GlobalTracer::get()->limited()) {
+                return dd_trace_forward_call();
+            }
+
             return MemcachedIntegration::traceCommandByKey($this, 'touchByKey', func_get_args());
         });
 
@@ -223,7 +340,8 @@ class MemcachedIntegration extends Integration
 
     public static function traceCommand($memcached, $command, $args)
     {
-        $scope = GlobalTracer::get()->startIntegrationScopeAndSpan(
+        $tracer = GlobalTracer::get();
+        $scope = $tracer->startIntegrationScopeAndSpan(
             MemcachedIntegration::getInstance(),
             "Memcached.$command"
         );
@@ -245,7 +363,8 @@ class MemcachedIntegration extends Integration
 
     public static function traceCommandByKey($memcached, $command, $args)
     {
-        $scope = GlobalTracer::get()->startIntegrationScopeAndSpan(
+        $tracer = GlobalTracer::get();
+        $scope = $tracer->startIntegrationScopeAndSpan(
             MemcachedIntegration::getInstance(),
             "Memcached.$command"
         );
@@ -266,7 +385,8 @@ class MemcachedIntegration extends Integration
 
     public static function traceCas($memcached, $args)
     {
-        $scope = GlobalTracer::get()->startIntegrationScopeAndSpan(
+        $tracer = GlobalTracer::get();
+        $scope = $tracer->startIntegrationScopeAndSpan(
             MemcachedIntegration::getInstance(),
             'Memcached.cas'
         );
@@ -285,7 +405,8 @@ class MemcachedIntegration extends Integration
 
     public static function traceCasByKey($memcached, $args)
     {
-        $scope = GlobalTracer::get()->startIntegrationScopeAndSpan(
+        $tracer = GlobalTracer::get();
+        $scope = $tracer->startIntegrationScopeAndSpan(
             MemcachedIntegration::getInstance(),
             'Memcached.casByKey'
         );
@@ -306,7 +427,8 @@ class MemcachedIntegration extends Integration
 
     public static function traceMulti($memcached, $command, $args)
     {
-        $scope = GlobalTracer::get()->startIntegrationScopeAndSpan(
+        $tracer = GlobalTracer::get();
+        $scope = $tracer->startIntegrationScopeAndSpan(
             MemcachedIntegration::getInstance(),
             "Memcached.$command"
         );
@@ -324,7 +446,8 @@ class MemcachedIntegration extends Integration
 
     public static function traceMultiByKey($memcached, $command, $args)
     {
-        $scope = GlobalTracer::get()->startIntegrationScopeAndSpan(
+        $tracer = GlobalTracer::get();
+        $scope = $tracer->startIntegrationScopeAndSpan(
             MemcachedIntegration::getInstance(),
             "Memcached.$command"
         );
